added bigInt to migrations

+++ migrations/20190718114314_geolocation.js
@@ -1,7 +1,7 @@
 exports.up = function(knex) {
   return knex.schema.createTable('geolocation', table => {
       table.increments('id').primary()
-      table.integer('timestamp')
+      table.bigInteger('timestamp')
       table.decimal('latitude', 9, 6)
       table.decimal('longitude', 9, 6)
       table.decimal('latitude_rounded', 7, 4)
